empiezo a implementar el fileSystem

+++ lib/stdbool.h
@@ -1,7 +1,10 @@
 #ifndef STDBOOL_H
 #define STDBOOL_H
 
-typedef int bool;
+// Headers extra
+#include <stdint.h>
+
+typedef uint8 bool;
 
 #define true 1
 #define false 0
